pyproject.toml
[project]
name = "derin-ogrenme"
version = "0.1.0"
description = "Deep Learning Project"
requires-python = ">=3.10"
dependencies = [
    "tensorflow==2.15.0",
    "numpy>=1.26.4",
    "keras==2.15.0",
    "tensorboard==2.15.2",
    "datasets>=2.18.0",
    "transformers>=4.38.1",
    "pillow>=10.2.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/derin_ogrenme"] 